Feature: add query params on searchScreen
